backend/requirements.txt
# CardTrade 後端依賴
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
google-generativeai==0.8.3
Pillow==11.1.0
aiofiles==24.1.0
